pyproject.toml
[tool.poetry]
name = "wechat-chatbot"
version = "0.1.0"
description = "chatbot for wechat"
authors = ["niyoh <zhrh120@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.9"
edgegpt = "^0.1.3"
pydantic = {extras = ["dotenv"], version = "^1.10.6"}
itchat-uos = "1.5.0.dev0"
structlog = "^22.3.0"
rich = "^13.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
